wip: prover crate structure
