+++ crates/evm-prover/src/prover/circuits/mod.rs
@@ -1,0 +1,2 @@
+pub mod block;
+pub mod message;
